work train

+++ 42/KrpSim/hdezier2.0/include/task_type.h
@@ -6,7 +6,7 @@
 /*   By: leeios <leeios@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/06/13 07:26:17 by leeios            #+#    #+#             */
-/*   Updated: 2016/06/17 18:39:47 by leeios           ###   ########.fr       */
+/*   Updated: 2016/06/18 17:12:32 by leeios           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,6 +25,6 @@ typedef std::unordered_map<std::string, const Task *>	t_tasks_link;
 typedef std::unordered_map<std::string, t_tasks_link>	t_tasks_link_by_res;
 typedef std::vector<std::string>						t_tasks_name;
 typedef std::pair<std::string, uint64_t>				t_task_number;
-typedef std::unordered_map<std::string, uint64_t>		t_task_pack;
+typedef std::unordered_map<std::string, double>			t_task_pack;
 
 #endif
